database setup and user query sql commands

+++ src/db/setup.cc
@@ -3,7 +3,6 @@
 
 using namespace drogon;
 
-orm::DbClientPtr create_db(std::string db_name, int pool) {
-
+orm::DbClientPtr create_db(std::string db_name, size_t pool) {
   return orm::DbClient::newSqlite3Client("filename=" + db_name, pool);
 }
